+++ zcash_client_sqlite/src/wallet/init/migrations.rs
@@ -320,7 +320,7 @@ mod tests {
 
     use crate::{
         testing::db::{test_clock, test_rng},
-        wallet::init::init_wallet_db_internal,
+        wallet::init::WalletMigrator,
         WalletDb,
     };
 
@@ -338,12 +338,10 @@ mod tests {
 
         let seed = [0xab; 32];
         assert_matches!(
-            init_wallet_db_internal(
-                &mut db_data,
-                Some(Secret::new(seed.to_vec())),
-                migrations,
-                false
-            ),
+            WalletMigrator::new()
+                .with_seed(Secret::new(seed.to_vec()))
+                .ignore_seed_relevance()
+                .init_or_migrate_to(&mut db_data, migrations),
             Ok(_)
         );
     }
@@ -377,12 +375,10 @@ mod tests {
         let mut prev_leaves: &[Uuid] = &[];
         for migrations in super::PUBLIC_MIGRATION_STATES {
             assert_matches!(
-                init_wallet_db_internal(
-                    &mut db_data,
-                    Some(Secret::new(seed.clone())),
-                    migrations,
-                    false
-                ),
+                WalletMigrator::new()
+                    .with_seed(Secret::new(seed.clone()))
+                    .ignore_seed_relevance()
+                    .init_or_migrate_to(&mut db_data, migrations),
                 Ok(_)
             );
 
@@ -399,7 +395,10 @@ mod tests {
         // Now check that we can migrate from the last public release to the current
         // migration state in this branch.
         assert_matches!(
-            init_wallet_db_internal(&mut db_data, Some(Secret::new(seed)), &[], false),
+            WalletMigrator::new()
+                .with_seed(Secret::new(seed))
+                .ignore_seed_relevance()
+                .init_or_migrate(&mut db_data),
             Ok(_)
         );
         // We don't ensure that the migration state changed, because it may not have.